requirements.txt
speedtest-cli
influxdb-client